refactor: introduce canonical molecule package
Ported-from: e19ee28bb156ce982da5a74e7fe136f074c2647d
Ported-from: 2e00ce0e3f2065a3815743af21fafbd94da01197
