+++ synplan/chem/molecule/__init__.py
@@ -1,0 +1,24 @@
+"""Canonical molecule-domain helpers used by SynPlanner."""
+
+from synplan.chem.molecule.io import (
+    standardize_building_blocks,
+    standardize_sdf_text,
+    standardize_smiles_batch,
+)
+from synplan.chem.molecule.precursor import Precursor, compose_precursors
+from synplan.chem.molecule.standardization import (
+    mol_from_smiles,
+    safe_canonicalization,
+    unite_molecules,
+)
+
+__all__ = [
+    "Precursor",
+    "compose_precursors",
+    "mol_from_smiles",
+    "safe_canonicalization",
+    "standardize_building_blocks",
+    "standardize_sdf_text",
+    "standardize_smiles_batch",
+    "unite_molecules",
+]
